GraphRAG/requirements.txt
# GraphRAG 核心依赖
graphrag>=0.3.0

# Web 界面
streamlit>=1.29.0

# OpenAI SDK
openai>=1.10.0

# 文档处理
python-docx>=1.1.0
pdfplumber>=0.10.0
PyPDF2>=3.0.0

# 环境变量管理
python-dotenv>=1.0.0

# Token 计数
tiktoken>=0.5.0

# 数据处理
pandas>=2.0.0
numpy>=1.24.0

# HTTP 请求
requests>=2.31.0
httpx>=0.25.0

# 日志
loguru>=0.7.0

# 图和社区检测
networkx>=3.0
python-louvain>=0.16
graspologic>=3.2.0
